feat: make spinning can animation its own component
